projects/shared/ai-customer-service/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Twilio Voice
twilio>=8.10.0

# Speech-to-Text
deepgram-sdk>=3.0.0

# Text-to-Speech
elevenlabs>=1.0.0

# LLM
anthropic>=0.18.0
openai>=1.10.0

# POS Integrations
httpx>=0.26.0

# Billing
stripe>=7.0.0

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
